__app__/dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile '.\dev-requirements.in'
#
appdirs==1.4.3            # via black
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via black, pytest
azure-functions==1.2.0    # via -r .\__app__/requirements.txt, -r .\dev-requirements.in
black==19.10b0            # via -r .\dev-requirements.in
click==7.1.1              # via black, pip-tools
colorama==0.4.3           # via pytest
importlib-metadata==1.6.0  # via importlib-resources, pluggy, pytest
importlib-resources==1.4.0  # via -r .\dev-requirements.in
more-itertools==8.2.0     # via pytest
packaging==20.3           # via pytest
pathspec==0.8.0           # via black
pip-tools==5.0.0          # via -r .\dev-requirements.in
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pyparsing==2.4.7          # via packaging
pytest==5.4.1             # via -r .\dev-requirements.in
regex==2020.4.4           # via black
six==1.14.0               # via packaging, pip-tools
toml==0.10.0              # via black
typed-ast==1.4.1          # via black
wcwidth==0.1.9            # via pytest
zipp==3.1.0               # via importlib-metadata, importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
